backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "API Backend server for LoL Coach"
dependencies = [
    "core",
    "ml",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.optional-dependencies]
genai = [
    "google-genai",
    "openai",
]
dev = [
    "google-genai",
    "openai",
    "httpx",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "types-PyYAML",
    "types-requests",
]

[tool.setuptools.packages.find]
where = ["src"]
